+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -134,6 +134,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-detail-prereq
+  <clang-tidy/checks/hooks-detail-prereq>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-entry-point-recursion
   <clang-tidy/checks/hooks-entry-point-recursion>` check.